requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.6.0
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
enzyme==0.4.1
idna==2.5                 # via requests
olefile==0.44             # via pillow
pillow==4.2.1
requests==2.18.2
urllib3==1.22             # via requests
